requirements.txt.txt
flaskpip install langchain-groq
langchain-groq
langchain-core
python-dotenv 